Rename module manager to managers
Corrected typo in module name, it should be `managers.py`
