+++ datahub/omis/invoice/models.py
@@ -6,7 +6,7 @@ from django.db import models
 from datahub.core.models import BaseModel
 from datahub.metadata.models import Country
 
-from .manager import InvoiceManager
+from .managers import InvoiceManager
 
 MAX_LENGTH = settings.CHAR_FIELD_MAX_LENGTH
