+++ datahub/omis/quote/managers.py
